+++ node/libs/roles/Cargo.toml
@@ -22,6 +22,9 @@ serde.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 
+[dev-dependencies]
+assert_matches.workspace = true
+
 [build-dependencies]
 zksync_protobuf_build.workspace = true
