+++ node/libs/roles/src/validator/messages/mod.rs
@@ -4,7 +4,15 @@ mod block;
 mod consensus;
 mod discovery;
 mod msg;
+mod replica_prepare;
+mod leader_prepare;
+mod replica_commit;
+mod leader_commit;
 
+pub use replica_prepare::*;
+pub use leader_prepare::*;
+pub use replica_commit::*;
+pub use leader_commit::*;
 pub use block::*;
 pub use consensus::*;
 pub use discovery::*;
